+++ examples/get-started/pure-js-google-maps/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "pure-js",
+  "name": "pure-js-google-maps",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
@@ -8,11 +8,11 @@
     "build": "NODE_ENV=production webpack --env.prod=true"
   },
   "dependencies": {
-    "@deck.gl/core": "^6.2.0",
-    "@deck.gl/layers": "^6.2.0"
+    "@deck.gl/core": "^7.0.0-beta",
+    "@deck.gl/google-maps": "^7.0.0-beta",
+    "@deck.gl/layers": "^7.0.0-beta"
   },
   "devDependencies": {
-    "html-webpack-plugin": "^3.0.7",
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.1"
