+++ examples/get-started/pure-js-mapbox/app.js
@@ -1,0 +1,56 @@
+import {Deck} from '@deck.gl/core';
+import {GeoJsonLayer} from '@deck.gl/layers';
+import mapboxgl from 'mapbox-gl';
+
+// Outlines of US States. Source: Natural Earth http://www.naturalearthdata.com/ via geojson.xyz
+const US_MAP_GEOJSON =
+  'https://d2ad6b4ur7yvpq.cloudfront.net/naturalearth-3.3.0/ne_110m_admin_1_states_provinces_shp.geojson'; //eslint-disable-line
+
+const INITIAL_VIEW_STATE = {
+  latitude: 40,
+  longitude: -100,
+  zoom: 3,
+  bearing: 30,
+  pitch: 30
+};
+
+// Set your mapbox token here
+mapboxgl.accessToken = process.env.MapboxAccessToken; // eslint-disable-line
+
+const map = new mapboxgl.Map({
+  container: 'map',
+  style: 'mapbox://styles/mapbox/light-v9',
+  // Note: deck.gl will be in charge of interaction and event handling
+  interactive: false,
+  center: [INITIAL_VIEW_STATE.longitude, INITIAL_VIEW_STATE.latitude],
+  zoom: INITIAL_VIEW_STATE.zoom,
+  bearing: INITIAL_VIEW_STATE.bearing,
+  pitch: INITIAL_VIEW_STATE.pitch
+});
+
+export const deck = new Deck({
+  canvas: 'deck-canvas',
+  width: '100%',
+  height: '100%',
+  initialViewState: INITIAL_VIEW_STATE,
+  controller: true,
+  onViewStateChange: ({viewState}) => {
+    map.jumpTo({
+      center: [viewState.longitude, viewState.latitude],
+      zoom: viewState.zoom,
+      bearing: viewState.bearing,
+      pitch: viewState.pitch
+    });
+  },
+  layers: [
+    new GeoJsonLayer({
+      data: US_MAP_GEOJSON,
+      stroked: true,
+      filled: true,
+      lineWidthMinPixels: 2,
+      opacity: 0.4,
+      getLineColor: [255, 100, 100],
+      getFillColor: [200, 160, 0, 180]
+    })
+  ]
+});
