+++ examples/get-started/pure-js-mapbox/index.html
