+++ examples/get-started/pure-js-mapbox/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "pure-js-without-map",
+  "name": "pure-js-mapbox",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
@@ -9,10 +9,10 @@
   },
   "dependencies": {
     "@deck.gl/core": "^6.2.0",
-    "@deck.gl/layers": "^6.2.0"
+    "@deck.gl/layers": "^6.2.0",
+    "mapbox-gl": "~0.44.2"
   },
   "devDependencies": {
-    "html-webpack-plugin": "^3.0.7",
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.1"
