+++ examples/get-started/pure-js/README.md
@@ -1,0 +1,12 @@
+<div align="center">
+   <img width="150" heigth="150" src="https://webpack.js.org/assets/icon-square-big.svg" />
+</div>
+
+## Example: Use deck.gl without React and Base Map
+
+Uses [Webpack](https://github.com/webpack/webpack) to bundle files and serves it
+with [webpack-dev-server](https://webpack.js.org/guides/development/#webpack-dev-server).
+
+Commands:
+* `npm start` is the development target, and will serves the app on port 3000.
+* `npm run build` is the production target, to create the final bundle.
